% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_surv_statistics.R
\name{plot_recurrent_surv_ci}
\alias{plot_recurrent_surv_ci}
\title{Plot recurrent-survival confidence intervals or bands}
\usage{
plot_recurrent_surv_ci(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link[=recurrent_surv_ci]{recurrent_surv_ci()}}.}

\item{...}{Passed to \code{\link[=plot]{plot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot recurrent-survival confidence intervals or bands
}
